Restructure project to follow domain-driven design
