+++ app/domain/permission/model.py
@@ -1,10 +1,19 @@
 from sqlalchemy import Column
+from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
 from sqlalchemy import String
+from sqlalchemy import Table
 from sqlalchemy.orm import relationship
 
-from app.models import Base
-from app.models._links import user_permission_table
+from app.core.database import Base
+
+# Association table for User and Permission
+user_permission_table = Table(
+    "user_permission",
+    Base.metadata,
+    Column("user_id", Integer, ForeignKey("user.id"), primary_key=True),
+    Column("permission_id", Integer, ForeignKey("permission.id"), primary_key=True),
+)
 
 
 class Permission(Base):
