+++ app/domain/post/enums.py
@@ -1,0 +1,9 @@
+import enum
+
+
+class PostStatus(enum.Enum):
+    DRAFT = "DRAFT"
+    PUBLISHED = "PUBLISHED"
+
+
+__all__ = ["PostStatus"]
