--- app/models/_links.py
@@ -1,14 +1,0 @@
-from sqlalchemy import Column
-from sqlalchemy import ForeignKey
-from sqlalchemy import Integer
-from sqlalchemy import Table
-
-from app.models import Base
-
-# Association table for User and Permission
-user_permission_table = Table(
-    "user_permission",
-    Base.metadata,
-    Column("user_id", Integer, ForeignKey("user.id"), primary_key=True),
-    Column("permission_id", Integer, ForeignKey("permission.id"), primary_key=True),
-)
